Se añade la edifición de usuarios, sólo para administrador. Es una tabla, que cualquier cambio que se realice, se guarda inmediatamente.

+++ src/main/java/com/example/WebLogin/persistence/repository/UserRepository.java
@@ -12,7 +12,7 @@ public interface UserRepository extends CrudRepository<UserEntity, Long>  {
 
     Optional<UserEntity> findUserEntityByUsername(String username);
 
-    UserEntity findUserEntityById(int id);
+    UserEntity findUserEntityById(Long id);
 
     UserEntity getUserByUsername(String username);
